+++ src/main/java/org/vetmeduni/io/readers/fastq/paired/FastqReaderPairedImpl.java
@@ -20,7 +20,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  */
-package org.vetmeduni.io.readers.paired;
+package org.vetmeduni.io.readers.fastq.paired;
 
 import htsjdk.samtools.SAMException;
 import htsjdk.samtools.fastq.FastqReader;