+++ src/main/java/org/vetmeduni/io/readers/fastq/paired/FastqReaderPairedInterface.java
@@ -20,10 +20,10 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  */
-package org.vetmeduni.io.readers.single;
+package org.vetmeduni.io.readers.fastq.paired;
 
-import htsjdk.samtools.fastq.FastqRecord;
-import org.vetmeduni.io.readers.FastqReaderInterface;
+import org.vetmeduni.io.FastqPairedRecord;
+import org.vetmeduni.io.readers.fastq.FastqReaderInterface;
 
 import java.util.Iterator;
 
@@ -33,6 +33,7 @@ import java.util.Iterator;
  *
  * @author Daniel Gómez-Sánchez
  */
-public interface FastqReaderSingleInterface extends Iterator<FastqRecord>, Iterable<FastqRecord>, FastqReaderInterface {
+public interface FastqReaderPairedInterface
+	extends Iterator<FastqPairedRecord>, Iterable<FastqPairedRecord>, FastqReaderInterface {
 
 }