+++ src/main/java/org/vetmeduni/io/readers/fastq/paired/FastqReaderPairedSanger.java
@@ -20,30 +20,51 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  */
-package org.vetmeduni.io.readers;
+package org.vetmeduni.io.readers.fastq.paired;
 
+import htsjdk.samtools.fastq.FastqReader;
 import htsjdk.samtools.util.FastqQualityFormat;
+import org.vetmeduni.io.FastqPairedRecord;
+import org.vetmeduni.utils.fastq.QualityUtils;
+import org.vetmeduni.utils.record.FastqRecordUtils;
 
-import java.io.Closeable;
+import java.io.File;
 
 /**
- * Interface for implement different FastqReaders pair-end or single-end and contains information about the encoding
+ * Implementation for pair-end reader with two files that always returns a Sanger encoded record
  *
  * @author Daniel Gómez-Sánchez
  */
-public interface FastqReaderInterface extends Closeable {
+public class FastqReaderPairedSanger extends FastqReaderPairedImpl implements FastqReaderPairedInterface {
+
+	public FastqReaderPairedSanger(FastqReader reader1, FastqReader reader2) throws QualityUtils.QualityException {
+		super(reader1, reader2);
+	}
+
+	public FastqReaderPairedSanger(File reader1, File reader2) throws QualityUtils.QualityException {
+		super(reader1, reader2);
+	}
 
 	/**
-	 * Get the FASTQ quality for the reads. All the records returned should be in this format
+	 * Next always return a Sanger formatted record
 	 *
-	 * @return the FastqQuality for this reader
+	 * @return the next record
 	 */
-	public FastqQualityFormat getFastqQuality();
+	@Override
+	public FastqPairedRecord next() {
+		if (QualityUtils.isStandard(encoding)) {
+			return super.next();
+		}
+		return FastqRecordUtils.copyToSanger(super.next());
+	}
 
 	/**
-	 * Get the original encoding for the file
+	 * The returning format is always Sanger
 	 *
-	 * @return the original FastqQuality (in the file)
+	 * @return {@link htsjdk.samtools.util.FastqQualityFormat#Standard}
 	 */
-	public FastqQualityFormat getOriginalEncoding();
+	@Override
+	public FastqQualityFormat getFastqQuality() {
+		return FastqQualityFormat.Standard;
+	}
 }